Refactoring to prevent future MSVC warnings: * i_mouse -> i_mouse_button * i_mouse_x/y ints -> i_mouse vector

+++ src/game/g_ship.c
@@ -570,7 +570,7 @@ static void ship_quick_info(int index)
         /* Crew */
         color = I_COLOR_ALT;
         i = ship->store.cargo[G_CT_CREW].amount;
-        total = ship->store.capacity * G_SHIP_OPTIMAL_CREW;
+        total = (int)(ship->store.capacity * G_SHIP_OPTIMAL_CREW);
         prop = (float)i / total;
         if (prop >= 0.67)
                 color = I_COLOR_GOOD;
